pyproject.toml
[project]
name = "agentic-dev"
version = "0.1.0"
description = "Trialling some of the newer agentic releases"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.29",
    "langchain[openai]>=0.3.27",
    "langsmith>=0.4.25",
    "openai>=1.106.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
